ml_service/requirements.txt
fastapi==0.119.0
uvicorn==0.22.0
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5
mysql-connector-python==9.0.0